Moved a ton of client-related stuff back under client.py
